Move dialogue to view
