+++ app/controllers/settings_controller.py
@@ -6,7 +6,7 @@ from loguru import logger
 from PySide6.QtCore import QObject, Slot
 from PySide6.QtWidgets import QApplication
 
-from app.models.dialogue import show_dialogue_confirmation, show_dialogue_file
+from app.views.dialogue import show_dialogue_confirmation, show_dialogue_file
 from app.models.settings import Instance, Settings
 from app.utils.constants import SortMethod
 from app.utils.event_bus import EventBus
