+++ app/utils/steam/webapi/wrapper.py
@@ -12,7 +12,7 @@ from requests import post as requests_post
 from requests.exceptions import JSONDecodeError
 from steam.webapi import WebAPI
 
-from app.models.dialogue import show_dialogue_input, show_warning
+from app.views.dialogue import show_dialogue_input, show_warning
 from app.utils.app_info import AppInfo
 from app.utils.constants import RIMWORLD_DLC_METADATA
 from app.utils.generic import chunks
