+++ app/views/main_content_panel.py
@@ -47,7 +47,7 @@ from PySide6.QtCore import (
 from PySide6.QtWidgets import QFrame, QHBoxLayout, QLabel
 from requests import get as requests_get
 
-import app.models.dialogue as dialogue
+import app.views.dialogue as dialogue
 import app.utils.constants as app_constants
 import app.utils.metadata as metadata
 from app.controllers.sort_controller import Sorter
