+++ app/views/main_window.py
@@ -27,7 +27,7 @@ from app.controllers.menu_bar_controller import MenuBarController
 from app.controllers.settings_controller import (
     SettingsController,
 )
-from app.models.dialogue import (
+from app.views.dialogue import (
     show_dialogue_conditional,
     show_dialogue_confirmation,
     show_dialogue_file,
