+++ app/views/mods_panel.py
@@ -38,7 +38,7 @@ from PySide6.QtWidgets import (
 )
 
 from app.controllers.settings_controller import SettingsController
-from app.models.dialogue import (
+from app.views.dialogue import (
     show_dialogue_conditional,
     show_dialogue_input,
     show_warning,
