+++ app/windows/rule_editor_panel.py
@@ -30,7 +30,7 @@ from PySide6.QtWidgets import (
     QWidget,
 )
 
-from app.models.dialogue import show_dialogue_input, show_warning
+from app.views.dialogue import show_dialogue_input, show_warning
 from app.utils.app_info import AppInfo
 from app.utils.metadata import MetadataManager
