+++ app/windows/runner_panel.py
@@ -16,7 +16,7 @@ from PySide6.QtWidgets import (
     QWidget,
 )
 
-from app.models.dialogue import show_dialogue_conditional, show_dialogue_file
+from app.views.dialogue import show_dialogue_conditional, show_dialogue_file
 from app.utils.app_info import AppInfo
 from app.utils.steam.webapi.wrapper import (
     ISteamRemoteStorage_GetPublishedFileDetails,
